feat(PM25): 修改PM2.5数值类型为浮点型以提高精度
将PM25.h中pm25_value的数据类型从uint16_t改为float，
以支持更高精度的PM2.5浓度计算与显示。

feat(sensor): 新增传感器数据采集模块

创建SensorData模块，统一管理DHT11温湿度、光照强度及PM2.5传感器的数据读取。
该模块通过FreeRTOS任务定时采集各传感器数据，并打印至串口调试信息。

+++ User/Hardware/PM25.h
@@ -13,7 +13,7 @@
 
 // PM2.5结构体
 typedef struct {
-    uint16_t pm25_value;      // PM2.5数值 (μg/m³)
+    float pm25_value;      // PM2.5数值 (μg/m³)
     uint8_t  level;          // 污染等级 (0-5)
     float    voltage;         // ADC电压值 (V)
     uint16_t adc_raw;        // ADC原始值
